Move combat autoresolve result
