+++ Assets/Scripts/Game/Results/SpaceCombatAutoResult.cs.meta
